Cleanup internal structure a bit
